+++ config.php
@@ -7,6 +7,7 @@
  */
 
 use humhub\components\Application;
+use humhub\components\gates\GateManager;
 use humhub\modules\admin\grid\UserActionColumn;
 use humhub\modules\twofa\Events;
 use humhub\modules\user\controllers\AuthController;
@@ -19,6 +20,7 @@ return [
     'namespace' => 'humhub\modules\twofa',
     'events' => [
         [Application::class, Application::EVENT_BEFORE_REQUEST, [Events::class, 'onBeforeRequest']],
+        [GateManager::class, GateManager::EVENT_INIT_GATES, [Events::class, 'onGateInit']],
         [AuthController::class, AuthController::EVENT_AFTER_LOGIN, [Events::class, 'onAfterLogin']],
         [Controller::class, Controller::EVENT_BEFORE_ACTION, [Events::class, 'onBeforeAction']],
         [Controller::class, Controller::EVENT_AFTER_ACTION, [Events::class, 'onAfterAction']],